Attempt to get a better overview of the pundit tests
The spec helper is large, contains a mixture of policies and entities and is hard to follow. By splitting it out into smaller chunks we can get a better overview.
