+++ pundit.gemspec
@@ -32,4 +32,5 @@ Gem::Specification.new do |gem|
   gem.add_development_dependency "rubocop"
   gem.add_development_dependency "simplecov", ">= 0.17.0"
   gem.add_development_dependency "yard"
+  gem.add_development_dependency "zeitwerk"
 end
