+++ spec/support/models/article.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Article
+end
