+++ spec/support/models/article_tag.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ArticleTag
+  def self.policy_class
+    ArticleTagOtherNamePolicy
+  end
+end
