+++ spec/support/models/artificial_blog.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class ArtificialBlog < Blog
+  def self.policy_class
+    BlogPolicy
+  end
+end
