+++ spec/support/models/blog.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Blog
+end
