+++ spec/support/models/comment.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Comment
+  extend ActiveModel::Naming
+end
