+++ spec/support/models/comment_four_five_six.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CommentFourFiveSix
+  extend ActiveModel::Naming
+end
