+++ spec/support/models/default_scope_contains_error.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class DefaultScopeContainsError
+  def self.all; end
+end
