+++ spec/support/models/foo.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Foo
+end
