+++ spec/support/models/post_four_five_six.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class PostFourFiveSix
+  def initialize(user)
+    @user = user
+  end
+
+  attr_reader(:user)
+end
