+++ spec/support/models/project_one_two_three/avatar_four_five_six.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ProjectOneTwoThree
+  class AvatarFourFiveSix
+    extend ActiveModel::Naming
+  end
+end
