+++ spec/support/models/project_one_two_three/tag_four_five_six.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module ProjectOneTwoThree
+  class TagFourFiveSix
+    def initialize(user)
+      @user = user
+    end
+
+    attr_reader(:user)
+  end
+end
