+++ spec/support/models/wiki.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class Wiki
+end
