+++ spec/support/policies/article_tag_other_name_policy.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class ArticleTagOtherNamePolicy < BasePolicy
+  def show?
+    true
+  end
+
+  def destroy?
+    false
+  end
+
+  alias tag record
+end
