+++ spec/support/policies/blog_policy.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class BlogPolicy < BasePolicy
+  alias blog record
+end
