+++ spec/support/policies/criteria_policy.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CriteriaPolicy < BasePolicy
+  alias criteria record
+end
