+++ spec/support/policies/denier_policy.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class DenierPolicy < BasePolicy
+  def update?
+    false
+  end
+end
