+++ spec/support/policies/nil_class_policy.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+class NilClassPolicy < BasePolicy
+  class Scope
+    def initialize(*)
+      raise Pundit::NotDefinedError, "Cannot scope NilClass"
+    end
+  end
+
+  def show?
+    false
+  end
+
+  def destroy?
+    false
+  end
+end
