+++ spec/support/policies/project/admin/comment_policy.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Project
+  module Admin
+    class CommentPolicy < BasePolicy
+      def update?
+        true
+      end
+
+      def destroy?
+        false
+      end
+    end
+  end
+end
