+++ spec/support/policies/project/criteria_policy.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Project
+  class CriteriaPolicy < BasePolicy
+    alias criteria record
+  end
+end
