+++ spec/support/policies/project_one_two_three/post_four_five_six_policy.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module ProjectOneTwoThree
+  class PostFourFiveSixPolicy < BasePolicy
+  end
+end
