+++ spec/support/policies/publication_policy.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class PublicationPolicy < BasePolicy
+  class Scope < BaseScope
+    def resolve
+      scope.published
+    end
+  end
+
+  def create?
+    true
+  end
+end
