+++ spec/support/policies/wiki_policy.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class WikiPolicy
+  class Scope
+    # deliberate typo method
+    def initalize; end
+  end
+end
